+++ apps/hubble/src/cli.ts
@@ -406,11 +406,16 @@ app
       throw ipMultiAddrResult.error;
     }
 
-    const bootstrapAddrs = (
-      (cliOptions.bootstrap ??
-        hubConfig.bootstrap ??
-        (network === FarcasterNetwork.MAINNET ? MAINNET_BOOTSTRAP_PEERS : [])) as string[]
-    )
+    let bootstrapList = (cliOptions.bootstrap ?? hubConfig.bootstrap ?? []) as string[];
+    if (network === FarcasterNetwork.MAINNET) {
+      // Add in all the mainnet bootstrap peers
+      bootstrapList.push(...MAINNET_BOOTSTRAP_PEERS);
+
+      // deduplicate
+      bootstrapList = Array.from(new Set(bootstrapList));
+    }
+
+    const bootstrapAddrs = bootstrapList
       .map((a) => parseAddress(a))
       .map((a) => {
         if (a.isErr()) {
@@ -516,6 +521,7 @@ app
       enableSnapshotToS3,
       s3SnapshotBucket: cliOptions.s3SnapshotBucket ?? hubConfig.s3SnapshotBucket,
       hubOperatorFid: parseInt(cliOptions.hubOperatorFid ?? hubConfig.hubOperatorFid),
+      connectToDbPeers: hubConfig.connectToDbPeers ?? true,
     };
 
     // Startup check for Hub Operator FID