+++ apps/hubble/src/network/sync/syncEngine.ts
@@ -72,7 +72,7 @@ const SYNC_INTERRUPT_TIMEOUT = 30 * 1000; // 30 seconds
 
 // A quick sync will only sync messages that are newer than 2 weeks.
 const QUICK_SYNC_PROBABILITY = 0.7; // 70% of the time, we'll do a quick sync
-export const QUICK_SYNC_TS_CUTOFF = 2 * 7 * 24 * 60 * 60; // 2 weeks, in seconds
+export const QUICK_SYNC_TS_CUTOFF = 2 * 24 * 60 * 60; // 2 days, in seconds
 
 const COMPACTION_THRESHOLD = 100_000; // Sync
 const BAD_PEER_BLOCK_TIMEOUT = 5 * 60 * 60 * 1000; // 5 hours, arbitrary, may need to be adjusted as network grows
@@ -741,7 +741,7 @@ class SyncEngine extends TypedEmitter<SyncEvents> {
             result.addResult(fnameResult);
 
             // And finally messages
-            const messagesResult = await this.fetchAndMergeMessages(missingMessageIds, rpcClient);
+            const messagesResult = await this.fetchAndMergeMessages(missingMessageIds, rpcClient, quickSyncCutoff);
             result.addResult(messagesResult);
 
             fullSyncResult.addResult(result);
@@ -941,18 +941,35 @@ class SyncEngine extends TypedEmitter<SyncEvents> {
     return new MergeResult(syncIds.length, promises.length, 0, syncIds.length - promises.length);
   }
 
-  public async fetchAndMergeMessages(syncIds: SyncId[], rpcClient: HubRpcClient): Promise<MergeResult> {
+  public async fetchAndMergeMessages(
+    syncIds: SyncId[],
+    rpcClient: HubRpcClient,
+    quickSyncCutoff: number,
+  ): Promise<MergeResult> {
     if (syncIds.length === 0) {
       return new MergeResult(); // empty merge result
     }
 
+    let filteredSyncIds = syncIds;
+    if (quickSyncCutoff > 0) {
+      const startSyncIDs = syncIds.length;
+      const exists = await Promise.all(syncIds.map((s) => this.trie.exists(s)));
+      filteredSyncIds = syncIds.filter((_, i) => !exists[i]);
+      log.info(
+        { startSyncIDs, filteredSyncIds: filteredSyncIds.length, peer: this._currentSyncStatus.peerId },
+        "Fetching messages for sync",
+      );
+    }
+
     let result = new MergeResult();
     const start = Date.now();
+
     const messagesResult = await rpcClient.getAllMessagesBySyncIds(
-      SyncIds.create({ syncIds: syncIds.map((s) => s.syncId()) }),
+      SyncIds.create({ syncIds: filteredSyncIds.map((s) => s.syncId()) }),
       new Metadata(),
       rpcDeadline(),
     );
+
     statsd().timing("syncengine.peer.get_all_messages_by_syncids_ms", Date.now() - start);
 
     await messagesResult.match(