+++ apps/hubble/src/storage/engine/index.ts
@@ -39,7 +39,7 @@ import {
   UserNameType,
   utf8StringToBytes,
   validations,
-  VerificationAddEthAddressMessage,
+  VerificationAddAddressMessage,
   VerificationRemoveMessage,
 } from "@farcaster/hub-nodejs";
 import { err, ok, ResultAsync } from "neverthrow";
@@ -589,7 +589,7 @@ class Engine extends TypedEmitter<EngineEvents> {
   /*                          Verification Store Methods                        */
   /* -------------------------------------------------------------------------- */
 
-  async getVerification(fid: number, address: Uint8Array): HubAsyncResult<VerificationAddEthAddressMessage> {
+  async getVerification(fid: number, address: Uint8Array): HubAsyncResult<VerificationAddAddressMessage> {
     const validatedFid = validations.validateFid(fid);
     if (validatedFid.isErr()) {
       return err(validatedFid.error);
@@ -606,7 +606,7 @@ class Engine extends TypedEmitter<EngineEvents> {
   async getVerificationsByFid(
     fid: number,
     pageOptions: PageOptions = {},
-  ): HubAsyncResult<MessagesPage<VerificationAddEthAddressMessage>> {
+  ): HubAsyncResult<MessagesPage<VerificationAddAddressMessage>> {
     const validatedFid = validations.validateFid(fid);
     if (validatedFid.isErr()) {
       return err(validatedFid.error);
@@ -621,7 +621,7 @@ class Engine extends TypedEmitter<EngineEvents> {
   async getAllVerificationMessagesByFid(
     fid: number,
     pageOptions: PageOptions = {},
-  ): HubAsyncResult<MessagesPage<VerificationAddEthAddressMessage | VerificationRemoveMessage>> {
+  ): HubAsyncResult<MessagesPage<VerificationAddAddressMessage | VerificationRemoveMessage>> {
     const validatedFid = validations.validateFid(fid);
     if (validatedFid.isErr()) {
       return err(validatedFid.error);