+++ packages/core/src/validations.ts
@@ -5,7 +5,7 @@ import { bytesCompare, bytesToUtf8String, utf8StringToBytes } from "./bytes";
 import { ed25519, eip712 } from "./crypto";
 import { HubAsyncResult, HubError, HubResult } from "./errors";
 import { getFarcasterTime, toFarcasterTime } from "./time";
-import { makeVerificationEthAddressClaim } from "./verifications";
+import { makeVerificationAddressClaim } from "./verifications";
 import { UserNameType } from "./protobufs";
 import { normalize } from "viem/ens";
 import { defaultPublicClients, PublicClients } from "./eth/clients";
@@ -135,11 +135,11 @@ export const validateFid = (fid?: number | null): HubResult<number> => {
 
 export const validateEthAddress = (address?: Uint8Array | null): HubResult<Uint8Array> => {
   if (!address || address.length === 0) {
-    return err(new HubError("bad_request.validation_failure", "address is missing"));
+    return err(new HubError("bad_request.validation_failure", "Ethereum address is missing"));
   }
 
   if (address.length !== 20) {
-    return err(new HubError("bad_request.validation_failure", "address must be 20 bytes"));
+    return err(new HubError("bad_request.validation_failure", "Ethereum address must be 20 bytes"));
   }
 
   return ok(address);
@@ -296,13 +296,10 @@ export const validateMessageData = async <T extends protobufs.MessageData>(
     bodyResult = validateLinkBody(data.linkBody);
   } else if (validType.value === protobufs.MessageType.USER_DATA_ADD && !!data.userDataBody) {
     bodyResult = validateUserDataAddBody(data.userDataBody);
-  } else if (
-    validType.value === protobufs.MessageType.VERIFICATION_ADD_ETH_ADDRESS &&
-    !!data.verificationAddEthAddressBody
-  ) {
+  } else if (validType.value === protobufs.MessageType.VERIFICATION_ADD_ADDRESS && !!data.verificationAddAddressBody) {
     // Special check for verification claim
     bodyResult = await validateVerificationAddEthAddressBody(
-      data.verificationAddEthAddressBody,
+      data.verificationAddAddressBody,
       validFid.value,
       validNetwork.value,
       publicClients,
@@ -325,23 +322,23 @@ export const validateMessageData = async <T extends protobufs.MessageData>(
 };
 
 export const validateVerificationAddEthAddressSignature = async (
-  body: protobufs.VerificationAddEthAddressBody,
+  body: protobufs.VerificationAddAddressBody,
   fid: number,
   network: protobufs.FarcasterNetwork,
   publicClients: PublicClients = defaultPublicClients,
 ): HubAsyncResult<Uint8Array> => {
-  if (body.ethSignature.length > 256) {
-    return err(new HubError("bad_request.validation_failure", "ethSignature > 256 bytes"));
+  if (body.claimSignature.length > 256) {
+    return err(new HubError("bad_request.validation_failure", "claimSignature > 256 bytes"));
   }
 
-  const reconstructedClaim = makeVerificationEthAddressClaim(fid, body.address, network, body.blockHash);
+  const reconstructedClaim = makeVerificationAddressClaim(fid, body.address, network, body.blockHash);
   if (reconstructedClaim.isErr()) {
     return err(reconstructedClaim.error);
   }
 
   const verificationResult = await eip712.verifyVerificationEthAddressClaimSignature(
     reconstructedClaim.value,
-    body.ethSignature,
+    body.claimSignature,
     body.address,
     body.verificationType,
     body.chainId,
@@ -353,10 +350,10 @@ export const validateVerificationAddEthAddressSignature = async (
   }
 
   if (!verificationResult.value) {
-    return err(new HubError("bad_request.validation_failure", "invalid ethSignature"));
+    return err(new HubError("bad_request.validation_failure", "invalid claimSignature"));
   }
 
-  return ok(body.ethSignature);
+  return ok(body.claimSignature);
 };
 
 export const validateUrl = (url: string): HubResult<string> => {
@@ -597,11 +594,11 @@ export const validateReactionBody = (body: protobufs.ReactionBody): HubResult<pr
 };
 
 export const validateVerificationAddEthAddressBody = async (
-  body: protobufs.VerificationAddEthAddressBody,
+  body: protobufs.VerificationAddAddressBody,
   fid: number,
   network: protobufs.FarcasterNetwork,
   publicClients: PublicClients,
-): HubAsyncResult<protobufs.VerificationAddEthAddressBody> => {
+): HubAsyncResult<protobufs.VerificationAddAddressBody> => {
   const validAddress = validateEthAddress(body.address);
   if (validAddress.isErr()) {
     return err(validAddress.error);
@@ -623,7 +620,12 @@ export const validateVerificationAddEthAddressBody = async (
 export const validateVerificationRemoveBody = (
   body: protobufs.VerificationRemoveBody,
 ): HubResult<protobufs.VerificationRemoveBody> => {
-  return validateEthAddress(body.address).map(() => body);
+  switch (body.protocol) {
+    case protobufs.Protocol.ETHEREUM:
+      return validateEthAddress(body.address).map(() => body);
+    default:
+      return err(new HubError("bad_request.validation_failure", "invalid verification protocol"));
+  }
 };
 
 export const validateUsernameProofBody = (